Generalized IndoorsScreen. Renamed to IndividualScreensTemplate. Populates with data from indoors or health category depending upon navigation.
